:recycle: Separate tokenizer code
