+++ src/container.c
@@ -1,0 +1,20 @@
+#include <stdarg.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include "9cc.h"
+
+void error(char *fmt, ...) {
+  va_list ap;
+  va_start(ap, fmt);
+  vfprintf(stderr, fmt, ap);
+  fprintf(stderr, "\n");
+  exit(1);
+}
+
+void error_at(char *loc, char *fmt, ...) {
+  int pos = loc - user_input;
+  fprintf(stderr, "%s\n", user_input);
+  fprintf(stderr, "%*s", pos, "");
+  fprintf(stderr, "^ ");
+  error(fmt);
+}
