+++ src/parser.c
@@ -1,26 +1,7 @@
-#include <ctype.h>
-#include <stdarg.h>
 #include <stdbool.h>
-#include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include "9cc.h"
-
-void error(char *fmt, ...) {
-  va_list ap;
-  va_start(ap, fmt);
-  vfprintf(stderr, fmt, ap);
-  fprintf(stderr, "\n");
-  exit(1);
-}
-
-void error_at(char *loc, char *fmt, ...) {
-  int pos = loc - user_input;
-  fprintf(stderr, "%s\n", user_input);
-  fprintf(stderr, "%*s", pos, "");
-  fprintf(stderr, "^ ");
-  error(fmt);
-}
 
 bool consume(char *op) {
   if (token->kind != TK_RESERVED || memcmp(token->str, op, token->len)) {
@@ -47,57 +28,6 @@ int expect_number() {
 }
 
 bool at_eof() { return token->kind == TK_EOF; }
-
-bool starts_with(char *str, char *op) {
-  return strncmp(str, op, strlen(op)) == 0;
-}
-
-Token *new_token(TokenKind kind, Token *cur, char *str, int len) {
-  Token *t = calloc(1, sizeof(Token));
-  t->kind = kind;
-  t->str = str;
-  t->len = len;
-  cur->next = t;
-  return t;
-}
-
-Token *tokenize(char *p) {
-  Token head;
-  head.next = NULL;
-  Token *cur = &head;
-
-  while (*p) {
-    if (isspace(*p)) {
-      p++;
-      continue;
-    }
-
-    if (starts_with(p, "==") || starts_with(p, "!=") || starts_with(p, "<=") ||
-        starts_with(p, ">=")) {
-      cur = new_token(TK_RESERVED, cur, p, 2);
-      p += 2;
-      continue;
-    }
-
-    if (strchr("+-*/()<>", *p)) {
-      cur = new_token(TK_RESERVED, cur, p++, 1);
-      continue;
-    }
-
-    if (isdigit(*p)) {
-      char *q = p;
-      cur = new_token(TK_NUM, cur, p, 0);
-      cur->val = strtol(p, &p, 10);
-      cur->len = p - q;
-      continue;
-    }
-
-    error("Failed to tokenize");
-  }
-
-  new_token(TK_EOF, cur, p, 0);
-  return head.next;
-}
 
 Node *new_node(NodeKind kind, Node *lhs, Node *rhs) {
   Node *node = calloc(1, sizeof(Node));
