+++ src/tokenizer.c
@@ -1,0 +1,56 @@
+#include <ctype.h>
+#include <stdbool.h>
+#include <stdlib.h>
+#include <string.h>
+#include "9cc.h"
+
+bool starts_with(char *str, char *op) {
+  return strncmp(str, op, strlen(op)) == 0;
+}
+
+Token *new_token(TokenKind kind, Token *cur, char *str, int len) {
+  Token *t = calloc(1, sizeof(Token));
+  t->kind = kind;
+  t->str = str;
+  t->len = len;
+  cur->next = t;
+  return t;
+}
+
+Token *tokenize(char *p) {
+  Token head;
+  head.next = NULL;
+  Token *cur = &head;
+
+  while (*p) {
+    if (isspace(*p)) {
+      p++;
+      continue;
+    }
+
+    if (starts_with(p, "==") || starts_with(p, "!=") || starts_with(p, "<=") ||
+        starts_with(p, ">=")) {
+      cur = new_token(TK_RESERVED, cur, p, 2);
+      p += 2;
+      continue;
+    }
+
+    if (strchr("+-*/()<>", *p)) {
+      cur = new_token(TK_RESERVED, cur, p++, 1);
+      continue;
+    }
+
+    if (isdigit(*p)) {
+      char *q = p;
+      cur = new_token(TK_NUM, cur, p, 0);
+      cur->val = strtol(p, &p, 10);
+      cur->len = p - q;
+      continue;
+    }
+
+    error("Failed to tokenize");
+  }
+
+  new_token(TK_EOF, cur, p, 0);
+  return head.next;
+}
